Moce isOutOfStock to utils
