+++ app/routes/_index/route.tsx
@@ -1,11 +1,10 @@
 import { LinksFunction, LoaderFunctionArgs } from '@remix-run/node';
 import { Link, MetaFunction, useLoaderData, useNavigate, json } from '@remix-run/react';
-import { products as wixStoresProducts } from '@wix/stores';
 import { getEcomApi } from '~/api/ecom-api';
 import { HeroImage } from '~/components/hero-image/hero-image';
 import { ProductCard } from '~/components/product-card/product-card';
 import { ROUTES } from '~/router/config';
-import { getUrlOriginWithPath } from '~/utils';
+import { getUrlOriginWithPath, isOutOfStock } from '~/utils';
 import styles from './index.module.scss';
 
 export const loader = async ({ request }: LoaderFunctionArgs) => {
@@ -43,9 +42,7 @@ export default function HomePage() {
                                 name={product.name}
                                 price={product.priceData ?? undefined}
                                 className={styles.productCard}
-                                outOfStock={
-                                    product.stock?.inventoryStatus === wixStoresProducts.InventoryStatus.OUT_OF_STOCK
-                                }
+                                outOfStock={isOutOfStock(product)}
                             />
                         </Link>
                     ) : null
