+++ package.json
@@ -26,6 +26,7 @@
     "@wix/sdk": "^1.12.12",
     "@wix/stores": "^1.0.172",
     "classnames": "^2.5.1",
+    "fast-deep-equal": "^3.1.3",
     "isbot": "^5.1.14",
     "js-cookie": "^3.0.5",
     "react": "^18.3.1",
